refactor to `document.ts`
